Flappy Bird_Nadine
this is my FlappyBird Game
you need the latest Unity Version 2019.1 

Game instructions: Use the touchpad to play the game
The goal is to collect points and not hit the objects.
To finish the game you have to press cmd q.
If you die during the game and want to start again, you have to click on the touchpad. The game will start again!

Have fun!